Fix filtering advanced lifelist.
It was caused by filter options conversion form Parameters to hash. I was calling #to_h on filter parameters. For basic lifelist it was performed
only once, on Parameters, producing HashWithIndifferentAccess. But for advanced list this HashWithIndifferentAccess was passed to child lists
and HashWithIndifferentAccess#to_h produced just Hash with string keys, effectively voiding the filter.